Add files via upload
[Real or Fake] : Fake Job Description Prediction
This dataset contains 18K job descriptions out of which about 800 are fake. The data consists of both textual information and meta-information about the jobs. The dataset can be used to create classification models which can learn the job descriptions which are fraudulent.

Inspiration
The dataset is very valuable as it can be used to answer the following questions:

Create a classification model that uses text data features and meta-features and predict which job description are fraudulent or real.
Identify key traits/features (words, entities, phrases) of job descriptions which are fraudulent in nature.
Run a contextual embedding model to identify the most similar job descriptions.
Perform Exploratory Data Analysis on the dataset to identify interesting insights from this dataset.